// examples/react/src/skins/frosted/types.ts
export interface FrostedSkinStyles {
  readonly MediaContainer: string;
  readonly Overlay: string;
  readonly Controls: string;
  readonly Button: string;
  readonly IconButton: string;
  readonly PlayButton: string;
  readonly PlayIcon: string;
  readonly PauseIcon: string;
  readonly PlayTooltipPopup: string;
  readonly PlayTooltip: string;
  readonly PauseTooltip: string;
  readonly TooltipPopup: string;
  readonly MuteButton: string;
  readonly VolumeHighIcon: string;
  readonly VolumeLowIcon: string;
  readonly VolumeOffIcon: string;
  readonly FullScreenButton: string;
  readonly FullScreenEnterIcon: string;
  readonly FullScreenExitIcon: string;
  readonly FullScreenTooltipPopup: string;
  readonly FullScreenEnterTooltip: string;
  readonly FullScreenExitTooltip: string;
  readonly SliderRoot: string;
  readonly SliderTrack: string;
  readonly SliderProgress: string;
  readonly SliderPointer: string;
  readonly SliderThumb: string;
  readonly PopoverPopup: string;
  readonly TimeControls: string;
  readonly TimeDisplay: string;
}